no boxing

+++ src/converter/one/convert/JfrToOtlp.java
@@ -161,7 +161,7 @@ public class JfrToOtlp extends JfrConverter {
 
         @Override
         public void visit(Event event, long samples, long value) {
-            long nanosFromStart = Double.valueOf((event.time - jfr.chunkStartTicks) / ticksPerNanosecond).longValue();
+            long nanosFromStart = (long) ((event.time - jfr.chunkStartTicks) / ticksPerNanosecond);
             long timeNanos = jfr.chunkStartNanos + nanosFromStart;
 
             Range range = idToRange.computeIfAbsent(event.stackTraceId, this::computeLocationRange);
